move tex2img tests to etcpak tests
